update HTMLSelector tests
